thermal: mediatek: Add tsensor support for MT67XX/MT68XX thermal system
The auxadc thermal system on MT67XX and MT68XX SoCs seems to sit in between
versions V1 and V2. The efuse layout is mostly V1, but it has additional
entries for the likes of ADC_OE which is only seen otherwise in the V2 layout.
The downstream driver's version of raw_to_mcelsius uses a modified formula
of the V2 one, but there is a VTS value for each sensor, slightly adjusted
values and slope correction. Add support for this.

Signed-off-by: bengris32 <bengris32@protonmail.ch>